jwt refactoring

+++ src/controllers/userController.js
@@ -5,7 +5,7 @@ const { validationResult } = require('express-validator');
 exports.getAll = async (req, res) => {
     try {
         const users = await User.findAll({ order: [['id', 'ASC']] });
-        send200(res, { users });
+        send200(res, { users, refresh: req.refresh });
     } catch (err) {
         send500(res);
     }
@@ -27,9 +27,9 @@ exports.create = async (req, res) => {
 
             const hashedPassword = await bcrypt.hash(password, 13);
             const newUser = { name, age, email, password: hashedPassword };
-            
-            const {dataValues} = await User.create(newUser);
-            send201(res, dataValues, `User ${dataValues.id} created successfully`);
+
+            const { dataValues } = await User.create(newUser);
+            send201(res, dataValues);
         }
     } catch (err) {
         send500(res);
@@ -40,7 +40,7 @@ exports.findById = async (req, res) => {
     try {
         const { id } = req.params;
         const user = await User.findOne({ where: { id } });
-        send200(res, { user });
+        send200(res, { user, refresh: req.refresh });
     } catch (err) {
         send500(res);
     }
@@ -57,7 +57,7 @@ exports.update = async (req, res) => {
             const { id } = req.params;
 
             await User.update({ name, age }, { where: { id } });
-            send200(res, { message: `User ${id} updated successfully` });
+            send200(res, { refresh: req.refresh });
         }
     } catch (err) {
         send500(res);
@@ -67,8 +67,14 @@ exports.update = async (req, res) => {
 exports.delete = async (req, res) => {
     try {
         const { id } = req.params;
-        await User.destroy({ where: { id } });
-        send200(res, { message: `User ${id} deleted successfully` });
+        const current = req.user.id;
+    
+        if (+id === +current) {
+            send400(res, null, 'User cannot delete himself');
+        } else {
+            await User.destroy({ where: { id } });
+            send200(res, { refresh: req.refresh });
+        }
     } catch (err) {
         send500(res);
     }
